Many refactorings. Extraneous code after refactoring was removed.
Removed all storage of difficulty from CharManager.

CharManager is now queried with "atDiffBorder" and "atPageBorder" to determine which direction, if any, to advance the curr_list_num (etc.)

It now has a function, "move_prot_by_border_checks(Bool, Bool)" which is used by the GameEngine to tell it whether the protagonist should *actually* flip from one side of the screen to the other (vert/horiz). This helps because of special cases like first/last page/difficulty.